Correction bug upload file
