+++ src/AppBundle/Resources/views/Contact/contact.html.twig
@@ -1,34 +1,92 @@
 {% extends 'layout.html.twig' %}
 {% set homepage = true %}
 {% block stylesheets %}
-
-    {{ parent() }}
-
-
-
-    {% stylesheets filter='cssrewrite, ?scssphp'
-
-        'bundles/app/css/contact.css' %}
-
-        <link rel="stylesheet" href="{{ asset_url }}" type="text/css">
-
-    {% endstylesheets %}
-
+{{ parent() }}
+{% stylesheets filter='cssrewrite, ?scssphp'
+	'//maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css'
+    '//code.jquery.com/ui/1.12.1/themes/base/jquery-ui.css'
+    'bundles/app/css/contact.css' %}
+    <link rel="stylesheet" href="{{ asset_url }}" type="text/css">
+{% endstylesheets %}
 {% endblock %}
-{% block body %}
+{% block body  %}
 
-<div class="container" id="containerContact">
-	{{ form_start(form) }}
-	{{ form_errors(form) }}
-	{{ form_row(form.name) }}
-	{{ form_row(form.email) }}
-	{{ form_row(form.subject) }}
-	{{ form_row(form.message) }}
-	{{ form_row(form.save) }}
-	{{ form_rest(form) }}
-	{{ form_end(form) }}
-{% endblock %}
+<div class="container" id="f">
+	<div class="row">
+		<div class="col-xs-12 col-sm-offset-1 col-sm-8" id="containerformContact">
+			{{ form_start(form) }}
+			{{ form_errors(form) }}
+			<div class="row field">
+				<div class="col-xs-12 col-col-sm-offset-1 col-sm-4">
+					{{ form_label(form.name) }}
+				</div>
+				<div class="col-xs-12 col-sm-offset-1 col-sm-4">
+					{{ form_widget(form.name) }}
+				</div>
+			</div>
+			<div class="row field">
+				<div class="col-xs-12 col-col-sm-offset-1 col-sm-4">
+					{{ form_label(form.email) }}
+				</div>
+				<div class="col-xs-12 col-sm-offset-1 col-sm-4">
+					{{ form_widget(form.email) }}
+				</div>
+			</div>
+			<div class="row field">
+				<div class="col-xs-12 col-col-sm-offset-1 col-sm-4">
+					{{ form_label(form.subject) }}
+				</div>
+				<div class="col-xs-12 col-sm-offset-1 col-sm-4">
+					{{ form_widget(form.subject) }}
+				</div>
+			</div>
+			<div class="row field">
+				<div class="col-xs-12 col-col-sm-offset-1 col-sm-4">
+					{{ form_label(form.message) }}
+				 {% image 'http://assets.fhuszti.tech/nao/logo.png' %}
+                <img class="img-responsive" src="{{ 'http://assets.fhuszti.tech/nao/pelican.png' }}" alt="{{'core.logo.alt'|trans}}" class="hidden-xs" />
+              {% endimage %}
+				</div>
+				<div class="col-xs-12 col-sm-offset-1 col-sm-4">
+					{{ form_widget(form.message) }}
+				</div>
+			</div>
+			<div id="save">
+				{{ form_row(form.save) }}
+			</div>
+			{{ form_rest(form) }}
+			{{ form_end(form) }}
+		</div>
+	</div>
+	<div class="row">
+		<div class="col-xs-12 col-sm-offset-1 col-sm-8" id="blocAdress">
+
+		</div>
+	</div>
+</div>
+
 {% block script %}
-<script>
-</script>
+<script type="text/javascript">
+    $("form").on("submit",function(e) {
+      e.preventDefault();
+      var form = $(this);
+      var values = {};
+console.log(values);
+      $.each( form.serializeArray(), function(i, field) { 
+       values[field.name] = field.value;   
+     });
+
+        $.ajax({
+          url : '{{ path('app_contact') }}',
+          type : 'POST', 
+          data : values,
+          success :function(e) {
+            $("#containerFormContact").hide(1000,function(){
+            $("#thanks").show(1000);
+            });
+          }
+        });
+      });
+    </script>
+{% endblock %}
 {% endblock %}
