+++ src/AppBundle/Resources/views/Observation/observation.html.twig
@@ -166,17 +166,15 @@
   <script type="text/javascript">
     $("form").on("submit",function(e) {
       e.preventDefault();
-      var form = $(this);
-      var values = {};
-
-      $.each( form.serializeArray(), function(i, field) { 
-       values[field.name] = field.value;   
-     });
+      var form = $('form')[0]; // You need to use standard javascript object here
+var formData = new FormData(form);
 
         $.ajax({
           url : '{{ path('app_observation') }}',
           type : 'POST', 
-          data : values,
+          data : formData,
+           contentType: false, // NEEDED, DON'T OMIT THIS (requires jQuery 1.6+)
+    processData: false,
           success :function(e) {
             $("#containerObservation").hide(1000,function(){
             $("#thanks").show(1000);
@@ -186,10 +184,11 @@
       });
     
     </script>
-<!--     <script>
+<!-- <script>
 $(document).ready(function(){
   $('#observation_birdName').keyup(function() {
     var value = $(this).val();
+    console.log(value);
       if(value === '' || value.length < 3) {
         $('#results').empty();
       }
@@ -198,31 +197,25 @@ $(document).ready(function(){
            url : '{{ path('search') }}', 
            type : 'GET', 
            data : 'search=' + value,
+
            success: function(reponse) { 
           $('#results').empty();
-            if (reponse.error) {
-              $('#results').prepend('<p>{{ 'core.find.msg.nofind'|trans }}</p>');
-            }
-            else {
+            // if (reponse.error) {
+            //   $('#results').prepend('<p>{{ 'core.find.msg.nofind'|trans }}</p>');
+            // }
+            // else {
           Object.keys(reponse).forEach(function(key) {
-              $('#results').prepend('<a href="'+ reponse[key].link +'" class="list-group-item theFind">'+ reponse[key].name +'</a>');
+              $('#results').prepend('<p class="list-group-item theFind">'+ reponse[key].name +'</a>');
           });
             }
-            },
-             error: function() { 
-                $('#results').empty();
-                $('#results').html('<p>{{ 'core.find.msg.error'|trans }}</p>');
-                }
+            // },
+             // error: function() { 
+             //    $('#results').empty();
+             //    $('#results').html('<p>{{ 'core.find.msg.error'|trans }}</p>');
+             //    }
         });
       }
   });
-});
-$("#results").on("click", "a.theFind", function(){
-  var href = $(this).attr('href');
-    $('#container_find').hide('slow');
-    $('#container_fiche').load(href);
-    $('#container_fiche').show('slow');
-    return false;
 });
 </script> -->
   </body>
